支持研究生课程表，支持自动更新周数

+++ app/src/main/java/com/dujiajun/courseblock/CourseActivity.java
@@ -13,37 +13,37 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
-import java.util.ArrayList;
+import com.dujiajun.courseblock.model.Course;
+
+import org.litepal.LitePal;
+
 import java.util.Arrays;
-import java.util.List;
 
 public class CourseActivity extends AppCompatActivity {
 
     public static final int ACTION_INSERT = 0;
     public static final int ACTION_MODIFY = 1;
     public static final int ACTION_DETAIL = 2;
-    private EditText etCourse, etTeacher, etLocation, etNote, etCourseId;
+    private final boolean[] isWeekSelected = new boolean[Course.MAX_WEEKS];
+    private final String[] weekItems = new String[Course.MAX_WEEKS];
+    private final String[] startItems = new String[Course.MAX_STEPS];
+    private EditText etCourse, etTeacher, etLocation, etNote, etCourseId, etClassId;
     private TextView tvWeeks, tvDay, tvStart, tvEnd;
     private int day;
     private int start;
     private int end;
     private int step;
-    private CourseManager courseManager;
-    private boolean[] isWeekSelected = new boolean[CourseManager.MAX_WEEKS];
-    private String[] weekItems = new String[CourseManager.MAX_WEEKS];
     private String[] dayItems;
-    private String[] startItems = new String[CourseManager.MAX_STEPS];
-    private String[] endItems;// = new String[CourseManager.MAX_STEPS - start + 1];
-    private List<Integer> weekList;
+    private String[] endItems;
+    private char[] weekCode = new char[Course.MAX_WEEKS];
     private int action;
-    private Course courseFromMain;
+    private Course course;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_course);
         initUI();
-        courseManager = CourseManager.getInstance(getApplicationContext());
     }
 
     private void initUI() {
@@ -56,6 +56,7 @@ public class CourseActivity extends AppCompatActivity {
         etLocation = findViewById(R.id.et_location);
         etNote = findViewById(R.id.et_note);
         etCourseId = findViewById(R.id.et_course_id);
+        etClassId = findViewById(R.id.et_class_id);
         tvWeeks = findViewById(R.id.tv_weeks);
         tvDay = findViewById(R.id.tv_day);
         tvDay.setText(getResources().getStringArray(R.array.days_in_week)[day]);
@@ -77,9 +78,9 @@ public class CourseActivity extends AppCompatActivity {
             end = start;
             step = 1;
             isWeekSelected[week - 1] = true;
-            weekList = new ArrayList<>();
+            course = new Course();
         } else if (action == ACTION_DETAIL) {
-            courseFromMain = (Course) intent.getSerializableExtra("course");
+            course = LitePal.find(Course.class, intent.getIntExtra("course", 0));
         }
         actionChange();
         refreshTextViewAfterDialog();
@@ -87,29 +88,30 @@ public class CourseActivity extends AppCompatActivity {
 
     private void loadInitData() {
         if (action == ACTION_DETAIL) {
-            etCourse.setText(courseFromMain.getCourseName());
-            etLocation.setText(courseFromMain.getLocation());
-            etTeacher.setText(courseFromMain.getTeacher());
-            etNote.setText(courseFromMain.getNote());
-            etCourseId.setText(courseFromMain.getCourseId());
-            day = courseFromMain.getDay();
-            start = courseFromMain.getStart();
-            step = courseFromMain.getStep();
+            etClassId.setText(course.getClassId());
+            etCourse.setText(course.getCourseName());
+            etLocation.setText(course.getLocation());
+            etTeacher.setText(course.getTeacher());
+            etNote.setText(course.getNote());
+            etCourseId.setText(course.getCourseId());
+            day = course.getDay();
+            start = course.getStart();
+            step = course.getStep();
             end = start + step - 1;
-            weekList = new ArrayList<>(courseFromMain.getWeekList());
-            for (int i = 0; i < isWeekSelected.length; i++) {
-                isWeekSelected[i] = weekList.contains(i + 1);
+            weekCode = course.getWeekCode().toCharArray();
+            for (int i = 0; i < weekCode.length; i++) {
+                isWeekSelected[i] = weekCode[i] == '1';
             }
         }
 
-        for (int i = 1; i <= CourseManager.MAX_WEEKS; i++) {
+        for (int i = 1; i <= Course.MAX_WEEKS; i++) {
             weekItems[i - 1] = String.format(getString(R.string.week), String.valueOf(i));
         }
         dayItems = getResources().getStringArray(R.array.days_in_week);
-        for (int i = 1; i <= CourseManager.MAX_STEPS; i++)
+        for (int i = 1; i <= Course.MAX_STEPS; i++)
             startItems[i - 1] = String.format(getString(R.string.period), String.valueOf(i));
-        endItems = new String[CourseManager.MAX_STEPS - start + 1];
-        for (int i = start; i <= CourseManager.MAX_STEPS; i++)
+        endItems = new String[Course.MAX_STEPS - start + 1];
+        for (int i = start; i <= Course.MAX_STEPS; i++)
             endItems[i - start] = String.format(getString(R.string.period), String.valueOf(i));
 
         tvStart.setText(startItems[start - 1]);
@@ -119,27 +121,22 @@ public class CourseActivity extends AppCompatActivity {
 
     private void refreshTextViewAfterDialog() {
         StringBuilder stringBuilder = new StringBuilder();
-        weekList.clear();
         for (int i = 0; i < isWeekSelected.length; i++) {
             if (isWeekSelected[i]) {
                 stringBuilder.append(i + 1);
                 stringBuilder.append(" ");
-                weekList.add(i + 1);
             }
         }
-
-        if (stringBuilder.length() != 0) {
-            stringBuilder.deleteCharAt(stringBuilder.length() - 1);
-            tvWeeks.setText(String.format(getString(R.string.week), stringBuilder.toString()));
-        } else {
-            tvWeeks.setText("");
-        }
+        tvWeeks.setText(stringBuilder.toString().trim());
     }
 
     private void showWeekDialog() {
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setTitle(getString(R.string.select_week));
-        builder.setMultiChoiceItems(weekItems, isWeekSelected, (dialog, which, isChecked) -> isWeekSelected[which] = isChecked);
+        builder.setMultiChoiceItems(weekItems, isWeekSelected, (dialog, which, isChecked) -> {
+            isWeekSelected[which] = isChecked;
+            weekCode[which] = isChecked ? '1' : '0';
+        });
         builder.setNeutralButton(R.string.ok,
                 (dialog, which) -> refreshTextViewAfterDialog());
         builder.setNegativeButton(getString(R.string.select_all), (dialog, which) -> {
@@ -161,10 +158,10 @@ public class CourseActivity extends AppCompatActivity {
         builder.setPositiveButton(R.string.ok,
                 (dialog, which) -> {
                     tvStart.setText(startItems[start - 1]);
-                    endItems = new String[CourseManager.MAX_STEPS - start + 1];
-                    for (int i = start; i <= CourseManager.MAX_STEPS; i++)
+                    endItems = new String[Course.MAX_STEPS - start + 1];
+                    for (int i = start; i <= Course.MAX_STEPS; i++)
                         endItems[i - start] = String.format(getString(R.string.period), String.valueOf(i));
-                    if (step - 1 < CourseManager.MAX_STEPS - start)
+                    if (step - 1 < Course.MAX_STEPS - start)
                         tvEnd.setText(endItems[step - 1]);
                     else
                         tvEnd.setText(endItems[0]);
@@ -186,8 +183,8 @@ public class CourseActivity extends AppCompatActivity {
     }
 
     private void showEndDialog() {
-        endItems = new String[CourseManager.MAX_STEPS - start + 1];
-        for (int i = start; i <= CourseManager.MAX_STEPS; i++)
+        endItems = new String[Course.MAX_STEPS - start + 1];
+        for (int i = start; i <= Course.MAX_STEPS; i++)
             endItems[i - start] = String.format(getString(R.string.period), String.valueOf(i));
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setTitle(getString(R.string.select_end));
@@ -203,58 +200,42 @@ public class CourseActivity extends AppCompatActivity {
     }
 
 
-    private Course checkAndNewCourse() {
+    private boolean checkCourse(boolean insert) {
         if ("".equals(etCourse.getText().toString().trim())) {
             Toast.makeText(this, getString(R.string.please_input_name), Toast.LENGTH_SHORT).show();
-            return null;
+            return false;
         }
         if ("".equals(tvWeeks.getText().toString().trim())) {
             Toast.makeText(this, getString(R.string.please_select_weeks), Toast.LENGTH_SHORT).show();
-            return null;
+            return false;
         }
         if (end < start) {
             Toast.makeText(this, getString(R.string.please_select_period), Toast.LENGTH_SHORT).show();
-            return null;
+            return false;
         }
-        Course course = new Course();
-        if (action == ACTION_MODIFY) {
-            course.setId(courseFromMain.getId());
-        }
+        course.setClassId(etClassId.getText().toString().trim());
         course.setCourseName(etCourse.getText().toString().trim());
         course.setTeacher(etTeacher.getText().toString().trim());
-        course.setWeekList(weekList);
+        course.setWeekCode(new String(weekCode));
         course.setStart(start);
         course.setStep(step);
         course.setDay(day);
         course.setLocation(etLocation.getText().toString().trim());
         course.setNote(etNote.getText().toString().trim());
         course.setCourseId(etCourseId.getText().toString().trim());
-        course.setFromServer(false);
-        return course;
-    }
-
-    private boolean checkAndInsert() {
-        Course course = checkAndNewCourse();
-        if (course == null)
-            return false;
-        courseManager.insertNewCourse(course);
+        course.setWeekCode(new String(weekCode));
+        course.setToDefault("isFromServer");
+        course.save();
         return true;
     }
 
-    private boolean checkAndModify() {
-        Course course = checkAndNewCourse();
-        if (course == null)
-            return false;
-        courseManager.modifyCourse(course);
-        return true;
-    }
 
     private void deleteCourse() {
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setTitle(getString(R.string.warning))
-                .setMessage(String.format(getString(R.string.sure_to_delete), courseFromMain.getCourseName()))
+                .setMessage(String.format(getString(R.string.sure_to_delete), course.getCourseName()))
                 .setPositiveButton(R.string.ok, (dialog, which) -> {
-                    courseManager.deleteCourseById(courseFromMain.getId());
+                    course.delete();
                     finish();
                 })
                 .setNegativeButton(R.string.cancel, null);
@@ -279,34 +260,28 @@ public class CourseActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_apply:
-                if (action == ACTION_INSERT) {
-                    if (checkAndInsert())
-                        finish();
-                } else if (action == ACTION_MODIFY) {
-                    if (checkAndModify())
-                        finish();
-                }
-                break;
-            case R.id.action_cancel:
-                action = ACTION_DETAIL;
-                invalidateOptionsMenu();
-                actionChange();
-                break;
-            case R.id.action_delete:
-                deleteCourse();
-                break;
-            case R.id.action_modify:
-                action = ACTION_MODIFY;
-                invalidateOptionsMenu();
-                actionChange();
-                break;
+        if (item.getItemId() == R.id.action_apply) {
+            if (action == ACTION_INSERT || action == ACTION_MODIFY) {
+                if (checkCourse(action == ACTION_INSERT))
+                    finish();
+            }
+        } else if (item.getItemId() == R.id.action_cancel) {
+            action = ACTION_DETAIL;
+            invalidateOptionsMenu();
+            actionChange();
+        } else if (item.getItemId() == R.id.action_delete) {
+            deleteCourse();
+        } else if (item.getItemId() == R.id.action_modify) {
+            action = ACTION_MODIFY;
+            invalidateOptionsMenu();
+            actionChange();
+
         }
         return true;
     }
 
     private void actionChange() {
+        etClassId.setEnabled(action != ACTION_DETAIL);
         etCourse.setEnabled(action != ACTION_DETAIL);
         etNote.setEnabled(action != ACTION_DETAIL);
         etTeacher.setEnabled(action != ACTION_DETAIL);
